refactor: revise db flavour setup
