+++ app/controllers/project_database.py
@@ -8,36 +8,7 @@ from app.helpers.database_service import MysqlDbService, PostgresqlDbService, ge
 from app.models.project import Project
 from flask_jwt_extended import jwt_required
 from app.helpers.decorators import admin_required
-
-# Used to return a database service
-
-database_flavours = [
-    {
-        'name': 'mysql',
-        'host': os.getenv('ADMIN_MYSQL_HOST'),
-        'port': os.getenv('ADMIN_MYSQL_PORT'),
-        'class': MysqlDbService()
-    },
-    {
-        'name': 'postgres',
-        'host': os.getenv('ADMIN_PSQL_HOST'),
-        'port': os.getenv('ADMIN_PSQL_PORT'),
-        'class': PostgresqlDbService()
-    }
-]
-
-
-# def get_db_service(flavour_name=None):
-#     pass
-
-
-def get_db_flavour(flavour_name=None):
-    if flavour_name == 'mysql':
-        return database_flavours[0]
-    elif flavour_name == 'postgres':
-        return database_flavours[1]
-    else:
-        return False
+from app.helpers.db_flavor import get_db_flavour
 
 
 class ProjectDatabaseView(Resource):
