+++ app/helpers/db_flavor.py
@@ -1,4 +1,5 @@
-
+from app.helpers.database_service import MysqlDbService, PostgresqlDbService
+import os
 
 db_flavors = {
     'postgres': {
@@ -17,3 +18,27 @@ db_flavors = {
         'port': 3306
     }
 }
+
+# Database flavours
+database_flavours = [
+    {
+        'name': 'mysql',
+        'host': os.getenv('ADMIN_MYSQL_HOST'),
+        'port': os.getenv('ADMIN_MYSQL_PORT'),
+        'class': MysqlDbService()
+    },
+    {
+        'name': 'postgres',
+        'host': os.getenv('ADMIN_PSQL_HOST'),
+        'port': os.getenv('ADMIN_PSQL_PORT'),
+        'class': PostgresqlDbService()
+    }
+]
+
+def get_db_flavour(flavour_name=None):
+    if flavour_name == 'mysql':
+        return database_flavours[0]
+    elif flavour_name == 'postgres':
+        return database_flavours[1]
+    else:
+        return False
